Refactor game imports
