+++ server/games/__init__.py
@@ -2,9 +2,15 @@ from collections import namedtuple
 
 from .coop import CoopGame
 from .custom_game import CustomGame
-from .game import Game
+from .game import Game, GameError
 from .ladder_game import LadderGame
-from .typedefs import FeaturedModType, GameState, VisibilityState
+from .typedefs import (
+    FeaturedModType,
+    GameState,
+    ValidityState,
+    Victory,
+    VisibilityState
+)
 
 FeaturedMod = namedtuple(
     "FeaturedMod",
@@ -15,10 +21,13 @@ FeaturedMod = namedtuple(
 __all__ = (
     "CoopGame",
     "CustomGame",
-    "Game",
-    "GameState",
-    "LadderGame",
     "FeaturedMod",
     "FeaturedModType",
-    "VisibilityState"
+    "Game",
+    "GameError",
+    "GameState",
+    "LadderGame",
+    "ValidityState",
+    "Victory",
+    "VisibilityState",
 )
