+++ tests/unit_tests/test_custom_game.py
@@ -2,8 +2,7 @@ import time
 
 import pytest
 
-from server.games import CustomGame
-from server.games.game import GameState, ValidityState
+from server.games import CustomGame, GameState, ValidityState
 from server.rating import RatingType
 from tests.unit_tests.conftest import add_connected_players
